refactor(migrator): INFRA-873 utils refactoring
