+++ packages/migrator/src/commands/addAppsKVPrefixes.ts
@@ -1,27 +1,12 @@
 import Table from 'cli-table3'
 
 import type { CommonOptions } from '@/utils/options'
-import type { PackageInfoWithLocation } from '@/utils/packages'
 
 import { extractEnvValue } from '@/utils/envs'
+import { isAppUsingKV } from '@/utils/kvdb/filtering'
+import { getKeyPrefix } from '@/utils/kvdb/keyPrefix'
 import { findApps, isNameMatching } from '@/utils/packages'
 
-const DEPS_TO_DETECT = ['ioredis', '@open-condo/keystone']
-
-function isAppUsingKV (pkg: PackageInfoWithLocation): boolean {
-    const devDeps = Object.keys(pkg.devDependencies || {})
-    const deps = Object.keys(pkg.dependencies || {})
-
-    const allDeps = [...devDeps, ...deps]
-
-    return DEPS_TO_DETECT.some(dep => allDeps.includes(dep))
-}
-
-function getPrefix (appName: string): string {
-    const scopedName = appName.split('/').pop() as string
-
-    return scopedName.replaceAll('-', '_').toLowerCase()
-}
 
 type TaskInfo = {
     appName: string
@@ -34,7 +19,7 @@ export async function addAppsKVPrefixes (options: CommonOptions): Promise<void> 
     const targetApps = allApps
         .filter(isAppUsingKV)
         .filter(pkg => isNameMatching(pkg, options.filter))
-        .toSorted()
+        .toSorted((a, b) => a.name.localeCompare(b.name))
 
     const connectionStrings = await Promise.all(targetApps.map(app => extractEnvValue(app, 'REDIS_URL')))
 
@@ -48,7 +33,7 @@ export async function addAppsKVPrefixes (options: CommonOptions): Promise<void> 
         .map(([appName, dbConnection]) => ({
             appName,
             dbConnection,
-            prefixKey: getPrefix(appName),
+            prefixKey: getKeyPrefix(appName),
         }))
 
     const table = new Table({
@@ -57,5 +42,5 @@ export async function addAppsKVPrefixes (options: CommonOptions): Promise<void> 
     table.push(...appsToPrefix.map(task => [task.appName, task.dbConnection, task.prefixKey]))
 
 
-    // console.log(appsToPrefix)
+    console.log(table.toString())
 }
