+++ packages/migrator/src/utils/kvdb/keyPrefix.ts
@@ -1,0 +1,5 @@
+export function getKeyPrefix (appName: string): string {
+    const scopedName = appName.split('/').pop() as string
+
+    return scopedName.replaceAll('-', '_').toLowerCase()
+}
